ref_block_prefix type changed to uint64

+++ src/FIO/Transaction.cpp
@@ -12,7 +12,7 @@ namespace TW::FIO {
 void Transaction::serialize(Data& out) const {
     encode32LE(expiration, out);
     encode16LE(refBlockNumber, out);
-    encode32LE(refBlockPrefix, out);
+    encode32LE(static_cast<uint32_t>(refBlockPrefix), out);
     encode32LE(0, out); // 00000000
     encodeVarInt(actions.size(), out);
     for (const auto& item : actions) {
